completed

+++ app/api/photo/[userID]/[photoID]/route.ts
@@ -1,18 +1,32 @@
 import { dbConfig } from "@/utils/dbConfig";
 import photoModel from "@/utils/model/labModel";
+import userModel from "@/utils/model/userModel";
+import { Types } from "mongoose";
 import { NextRequest, NextResponse } from "next/server";
 
 export const DELETE = async (req: NextRequest, { params }: any) => {
   try {
     await dbConfig();
-    const { photoID } = params;
+    const { photoID, userID } = params;
 
-    await photoModel.findByIdAndDelete(photoID);
+    const user = await userModel.findById(userID);
 
-    return NextResponse.json({
-      status: 200,
-      message: "Deleting personal images",
-    });
+    if (user) {
+      await photoModel.findByIdAndDelete(photoID);
+
+      user?.lab?.pull(new Types.ObjectId(photoID));
+      user?.save();
+
+      return NextResponse.json({
+        status: 201,
+        message: "Deleting personal images",
+      });
+    } else {
+      return NextResponse.json({
+        status: 404,
+        message: "Error getting user",
+      });
+    }
   } catch (error) {
     return NextResponse.json({
       status: 404,
